Substitute CurrentDispatcher.current for setDispatcher and move resolveDispatcher to CurrentDispathcer
