+++ src/CurrentDispatcher.js
@@ -2,4 +2,14 @@ let CurrentDispatcher = {
 	current: null,
 };
 
-export default CurrentDispatcher;
+export function setDispatcher(dispatcher) {
+	CurrentDispatcher.current = dispatcher;
+}
+
+export function resolveDispatcher() {
+	if (CurrentDispatcher.current) {
+		return CurrentDispatcher.current;
+	} else {
+		throw new Error('Express hooks should be used inside of express handlers or middlewares');
+	}
+}
