+++ src/__tests__/hooks.test.js
@@ -1,5 +1,5 @@
 import { useParam, useRes, useReq } from '../Hooks';
-import CurrentDispatcher from '../CurrentDispatcher';
+import { setDispatcher } from '../CurrentDispatcher';
 
 describe('hooks', () => {
 	it('hook throws error when used without dispatcher', () => {
@@ -8,18 +8,18 @@ describe('hooks', () => {
 
 	describe('useRes', () => {
 		it('Returns the response object correctly', () => {
-			CurrentDispatcher.current = {
+			setDispatcher({
 				_res: 'Res',
-			};
+			});
 			expect(useRes()).toMatch('Res');
 		});
 	});
 
 	describe('useReq', () => {
 		it('Returns the request object correctly', () => {
-			CurrentDispatcher.current = {
+			setDispatcher({
 				_req: 'Req',
-			};
+			});
 			expect(useReq()).toMatch('Req');
 		});
 	});
